feature(serve-static-pages):serve the UI static pages to the users browser
- serve the UI static pages to the users browser
- implement seperation of concern on the test script by splitting the single test file into 3 files for users, cars and orders
[Start #166328672]
