+++ package.json
@@ -50,6 +50,7 @@
   "dependencies": {
     "bcrypt": "^3.0.6",
     "body-parser": "^1.19.0",
+    "compression": "^1.7.4",
     "config": "^3.1.0",
     "coveralls": "^3.0.3",
     "debug": "^4.1.1",
@@ -57,7 +58,8 @@
     "eslint-plugin-import": "^2.17.2",
     "express": "^4.17.0",
     "istanbul": "^0.4.5",
-    "morgan": "^1.9.1"
+    "morgan": "^1.9.1",
+    "path": "^0.12.7"
   },
   "babel": {
     "presets": [
