+++ tests/carsTest.js
@@ -1,97 +1,6 @@
 import chai, { expect } from 'chai';
-import chaiHttp from 'chai-http';
-import config from 'config';
 
 import app from '../api/v1/index';
-import util from '../api/v1/util';
-
-chai.use(chaiHttp);
-chai.should();
-// testing to ensure the server is running
-describe('Server', () => {
-	it(`it should test that server is running on port ${config.get('port')}`, () => {
-		app.port.should.be.eql(config.get('port'));
-	});
-
-	it('it should send a message on default', () => {
-		chai.request(app)
-		.get('/')
-		.end((err, res) => {
-			expect(res.status).to.equal(200);
-			expect(res.body.message).to.equal('Welcome on Board: AutoMart API.');
-			expect(res.body).to.be.an.instanceof(Object);
-		});
-	});
-});
-
-// testing the user sign up and sign in endpoints
-describe('Testing User endpoints', () => {
-	const user = {
-		first_name: 'ChuksJoe',
-		last_name: 'Orjiakor',
-		email: 'chuksjoe@live.com',
-		password: 'testing',
-		is_admin: false,
-		address: {
-			street: '15 Aborishade road, Lawanson',
-			city: 'Surulere',
-			state: 'Lagos',
-			country: 'Nigeria',
-		},
-		phone: '08131172617',
-		zip: '234-001',
-	};
-	it('It should create new user account when valid entries are supplied', (done) => {
-		chai.request(app)
-		.post('/api/v1/auth/signup').set('Accept', 'application/json').send(user)
-		.end((err, res) => {
-      expect(res).to.have.status(201);
-      expect(res.body.data).to.include({
-        id: res.body.data.id,
-        token: res.body.data.token,
-        email: user.email,
-        first_name: user.first_name,
-      });
-      done();
-    });
-	});
-	it('It should not create new user account if any of the required entries are not supplied', (done) => {
-		// required entries: first_name, last_name, email, password, is_admin
-		user.first_name = undefined;
-		user.email = undefined;
-		chai.request(app)
-		.post('/api/v1/auth/signup').set('Accept', 'application/json').send(user)
-		.end((err, res) => {
-      expect(res).to.have.status(400);
-      expect(res.body.data).to.equal('One of the main entries is not defined.');
-      done();
-    });
-	});
-	it('it should allow a user to sign into their account if they supply valid credentials', (done) => {
-		chai.request(app)
-		.post('/api/v1/auth/signin').type('form').send({ email: 'chuksjoe@live.com', password: 'testing' })
-		.end((err, res) => {
-			const { data } = res.body;
-			res.should.have.status(201);
-			expect(data).to.include({
-          id: data.id,
-          token: data.token,
-          email: 'chuksjoe@live.com',
-          first_name: 'ChuksJoe',
-        });
-			done();
-		});
-	});
-	it('it should not allow a user to sign in if they supply invalid credentials', (done) => {
-		chai.request(app)
-		.post('/api/v1/auth/signin').type('form').send({ email: 'chuksjoe@live.com', password: 'wrongpassword' })
-		.end((err, res) => {
-			res.should.have.status(200);
-			expect(res.body.data).to.equal('Invalid Username or Password!');
-			done();
-		});
-	});
-});
 
 // testing the car ad endpoints
 describe('Testing the car sale ads endpoints', () => {
@@ -341,76 +250,5 @@ describe('Testing the car sale ads endpoints', () => {
 			expect(data).to.equal('Unauthorized Access!');
 		});
 		done();
-	});
-});
-
-describe('Tests for the orders api endpoints', () => {
-	it('it should allow a registered user to place a purchase order for an unsold car ad', (done) => {
-		chai.request(app)
-		.post('/api/v1/order').send({ car_id: 1, buyer_id: 2, price_offered: 1400000 })
-		.end((err, res) => {
-			const { data } = res.body;
-			res.should.have.status(201);
-			expect(data).to.include({
-				car_id: 1,
-				price_offered: 1400000,
-				buyer_id: 2,
-			});
-			done();
-		});
-	});
-	// user with id = 6 does not exist
-	it('it should not allow an unregistered user to place a purchase order', (done) => {
-		chai.request(app)
-		.post('/api/v1/order').send({ car_id: 5, buyer_id: 6, price_offered: 1400000 })
-		.end((err, res) => {
-			res.should.have.status(401);
-			expect(res.body.data).to.equal('Unauthorized Access!');
-			done();
-		});
-	});
-	// car with id = 2 is sold
-	it('it should not allow a registered user to place a purchase order for a sold car ad', (done) => {
-		chai.request(app)
-		.post('/api/v1/order').send({ car_id: 2, buyer_id: 2, price_offered: 1400000 })
-		.end((err, res) => {
-			res.should.have.status(401);
-			expect(res.body.data).to.equal('Unauthorized Access!');
-			done();
-		});
-	});
-	// car with id = 10 does not exist
-	it('it should not allow a registered user to place a purchase order for a car ad that does not exist', (done) => {
-		chai.request(app)
-		.post('/api/v1/order').send({ car_id: 8, buyer_id: 2, price_offered: 1400000 })
-		.end((err, res) => {
-			res.should.have.status(404);
-			expect(res.body.data).to.equal('Car does not exist!');
-			done();
-		});
-	});
-	it('it should allow a buyer to update the price he/she offered for a posted ad if its still pending', (done) => {
-		chai.request(app)
-		.patch('/api/v1/order/1/price').send({ buyer_id: 2, new_price: 1500000 })
-		.end((err, res) => {
-			const { data } = res.body;
-			res.should.have.status(201);
-			expect(data).to.include({
-				old_price_offered: data.old_price_offered,
-				new_price_offered: 1500000,
-				buyer_id: 2,
-			});
-			done();
-		});
-	});
-	it('it should not allow a user who is not the buyer to update the price offered', (done) => {
-		chai.request(app)
-		.patch('/api/v1/order/1/price').send({ buyer_id: 1, new_price: 1600000 })
-		.end((err, res) => {
-			const { data } = res.body;
-			res.should.have.status(401);
-			expect(data).to.equal('Unauthorized Access!');
-			done();
-		});
 	});
 });
